Completed looks of the pages. Changed file system design
Separated the files into the files themselves + their specific CSS files and then actual Components. Finished completing the looks for all the pages, though the navBar needs to be changed to what is inside the pages themselves, remove the navBar stuff from the pages and replace it with a call to navBar.

CSS for LS, AC pages of the NavBar elements aren't matching other pages. The position settings are incorrect for them. Otherwise, the entire site matches verbatim what is from Axure.