Worked on the functionality of Logging into the site
Close to getting the Login function working, does still throw a 400 Bad Request error. Currently, it's throwing an error due to userId being of type Integer in the backend and the input being of type String from the frontend. Need to set userId to type String for the moment and see if it works.