Fixed the CORS policy error; Still unable to return a user from backend.
Added CORS policy security configurations into the backend, was able to successfully call the getUser method in the backend from the frontend. Issue currently is that the Optional<CMUser> does not return a user even if a user is present in the database. FindByUserId is not returning a User at all even though it is positioned in CMUserRepository.

If this issue is fixed, I believe it will call back correctly and validate correctly. It will throw an error in the frontend if the Json query returns nothing.