Login Functionality completed; You can now login with credentials
Validation with Password/username; Frontend handles the Password validation and backend handles username validation.

Configure files added to handle CORS policy security issues, and added code to automatically navigate to the /Entry page after the Submit button is clicked and validation is done correctly. Navigation does not go off if the Password is incorrect.

To handle validation visibly, see the If block comparing Passwords. As for showing incorrect username, perhaps have an additional check somewhere to trigger if a user by a specific username is not found.