Added functionality for changeRequests in frontend and backend
Added methods for Get/Save for changeRequests in the backend, added the bare frame for calling a changeRequest in the frontend for ListScreenApp page, though it is still only displaying the last element in the array. Checks need to be made to be sure it's only displaying the last element and not overlapping them on the page.