Worked on the buttons functionality in ListScreenApp and sorting CRs
Added the ability to update the stateLevel of a change request depending on what button is pushed. Restricted changerequests depending on what role the user is. Added global object files that will need to be converted to document.cookies if a reload of the page is made at any point (which will need to be done for showing buttons being pushed unless we can force a re-render on changes to the page in react). Sorting change requests by id, and started work on the sorting ability on stateLevel, though it is not able to be done yet.