Finished functionality for ListScreen App
OperationUsers can now sort the list by Order and by ChangeId. Other users cannot sort manually. Buttons will now accurately display, with a owner of a changeRequest seeing all their changerequests, but can only unfreeze their own that are frozen or above that state. OperationUsers can fully Approve a changeRequest, being prompted a calendar date which is checked to ensure it is between the start and stop window dates before patching that info to the changeRequest and archiving it. Denying a changeRequest should archive it entirely as well.

Also did some cleanup on formatting in all the pages, fixing misc. whitespace and spacing for functions.